+++ products/src/main/java/com/globant/http/v1/HealthController.java
@@ -1,4 +1,4 @@
-package com.globant.http;
+package com.globant.http.v1;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
