+++ products/src/main/java/com/globant/http/v1/ProductsController.java
@@ -1,4 +1,4 @@
-package com.globant.http;
+package com.globant.http.v1;
 
 import domain.combo.Combo;
 import domain.combo.CreateCombo;
